{"schema":"decodex/commit/1","summary":"Split MCP prompts","authority":"XY-1115"}
